refactored RSVP form to split logic
